Update COR_sentence_events.sql
Corrections costs data used to be a single value applied to all years in the previous SIAL version. However, we have received yearly costs from Corrections with SIAL version 1.1, and this needs an additional date based join in the SIAL_COR_sentence_events script. This "correction" has been applied to the code.